requirements.txt
pydeck==0.4.0
requests==2.24.0
streamlit==0.64.0
numpy==1.18.4
fpdf==1.7.2
pandas==1.0.4
Pillow==7.2.0
python-dotenv==0.14.0
